# openapi.yaml
openapi: 3.0.3
info:
  title: HR Enterprise Integration (EAI) API
  version: 2.0.0
  description: |
    Dokumentasi lengkap API untuk sistem integrasi HR Enterprise yang terdiri dari
    **HRIS Service**, **Attendance Service**, **Payroll Service**, dan **Integration Service**.

    ## Enterprise Integration Patterns yang Diimplementasi
    - **Message Channel**: Queue RabbitMQ (`integration_queue`, `attendance_queue`, `payroll_queue`, `integration_dlq`)
    - **Message Translator**: Transformasi XML (Attendance) → Canonical JSON di `integration-service/translator.js`
    - **Content-Based Router**: Routing berdasarkan `eventType` di `integration-service/router.js`
    - **Message Endpoint / Adapter**: Koneksi setiap service ke RabbitMQ channel
    - **Dead Letter Channel**: DLQ `integration_dlq` via `integration_dlx` exchange untuk pesan gagal
    - **Canonical Data Model**: Format JSON standar internal (lihat schema `CanonicalMessage`)

    ## Format Data per Service
    | Service | Format Outbound ke RabbitMQ |
    |---|---|
    | HRIS Service | **JSON** (`EmployeeCreated`) |
    | Attendance Service | **XML** (`AttendanceSubmitted`) |
    | Integration Service | **Canonical JSON** (output setelah translasi) |
    | Payroll Service | Konsumen Canonical JSON |
...
paths:
...
  /employees:
    servers:
      - url: http://localhost:3001
        description: HRIS Service
    get:
      tags: [HRIS]
      operationId: hris-getEmployees
      summary: Daftar semua karyawan
      description: Mengembalikan seluruh data karyawan yang tersimpan di database HRIS.
      responses:
        '200':
          description: Daftar karyawan berhasil diambil.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
              example:
                - id: "EMP-171758"
                  name: "Budi Santoso"
                  email: "budi@enterprise.com"
                  role: "Software Engineer"
                  basicSalary: 8000000
                  published: true
                  createdAt: "2026-06-05T10:00:00.000Z"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [HRIS]
      operationId: hris-createEmployee
      summary: Tambah karyawan baru
      description: |
        Membuat record karyawan baru di database lokal HRIS dengan status `published: false`.
        Karyawan harus di-publish terlebih dahulu agar tersinkronisasi ke service lain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
            example:
              name: "Budi Santoso"
              email: "budi@enterprise.com"
              role: "Software Engineer"
              basicSalary: 8000000
      responses:
        '201':
          description: Karyawan berhasil ditambahkan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
              example:
                id: "EMP-171758"
                name: "Budi Santoso"
                email: "budi@enterprise.com"
                role: "Software Engineer"
                basicSalary: 8000000
                published: false
                createdAt: "2026-06-05T10:00:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /payroll:
    servers:
      - url: http://localhost:3003
        description: Payroll Service
    get:
      tags: [Payroll]
      operationId: payroll-getAll
      summary: Daftar seluruh laporan payroll
      description: Mengembalikan semua record slip gaji yang telah digenerate oleh Payroll Service.
      responses:
        '200':
          description: Daftar payroll berhasil diambil.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payroll'
              example:
                - id: "PAY-192837"
                  employeeId: "EMP-171758"
                  employeeName: "Budi Santoso"
                  attendanceDays: 18
                  basicSalary: 8000000
                  deduction: 800000
                  totalSalary: 7200000
                  generatedAt: "2026-06-05T15:20:00.000Z"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payroll/generate:
    servers:
      - url: http://localhost:3003
        description: Payroll Service
    post:
      tags: [Payroll]
      operationId: payroll-generate
      summary: Generate payroll untuk semua karyawan
      description: |
        Menghitung gaji semua karyawan berdasarkan data kehadiran yang telah disinkronkan.

        **Aturan Bisnis:**
        - Standar hari kerja = 20 hari/bulan
        - Jika kehadiran < 20 hari: `Potongan = (20 - Kehadiran) × (Gaji Pokok / 20)`
        - Gaji bersih = `Gaji Pokok - Potongan` (minimal 0)
      responses:
        '200':
          description: Payroll berhasil digenerate untuk semua karyawan.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  payrolls:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payroll'
        '400':
          description: Tidak ada karyawan terdaftar di Payroll Service.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Tidak ada karyawan untuk dihitung payroll-nya."
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─────────────────────────────────────────────────
  # INTEGRATION MONITORING SERVICE  (http://localhost:3004)
  # ─────────────────────────────────────────────────
...
components:
  schemas:
...
    EmployeeInput:
      type: object
      required: [name, email, role, basicSalary]
      properties:
        name:
          type: string
          description: Nama lengkap karyawan
          example: "Budi Santoso"
        email:
          type: string
          format: email
          description: Alamat email unik karyawan
          example: "budi@enterprise.com"
        role:
          type: string
          description: Jabatan atau posisi karyawan
          example: "Software Engineer"
        basicSalary:
          type: number
          description: Gaji pokok per bulan (dalam Rupiah)
          example: 8000000
...
    Employee:
      type: object
      properties:
        id:
          type: string
          description: ID unik karyawan (auto-generated, format EMP-XXXXXX)
          example: "EMP-171758"
        name:
          type: string
          example: "Budi Santoso"
        email:
          type: string
          format: email
          example: "budi@enterprise.com"
        role:
          type: string
          example: "Software Engineer"
        basicSalary:
          type: number
          example: 8000000
        published:
          type: boolean
          description: Status apakah event EmployeeCreated sudah dikirim ke RabbitMQ
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-06-05T10:00:00.000Z"
...
    Payroll:
      type: object
      properties:
        id:
          type: string
          description: ID unik slip gaji (auto-generated, format PAY-XXXXXX)
          example: "PAY-192837"
        employeeId:
          type: string
          example: "EMP-171758"
        employeeName:
          type: string
          example: "Budi Santoso"
        attendanceDays:
          type: integer
          description: Total hari hadir yang terekam bulan ini
          example: 18
        basicSalary:
          type: number
          example: 8000000
        deduction:
          type: number
          description: "Potongan = (20 - attendanceDays) × (basicSalary / 20)"
          example: 800000
        totalSalary:
          type: number
          description: Gaji bersih setelah potongan
          example: 7200000
        generatedAt:
          type: string
          format: date-time
          example: "2026-06-05T15:20:00.000Z"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Pesan error dari server"
  responses:
    BadRequest:
      description: Request tidak valid (field wajib kosong atau format salah).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "All fields are required"
...
    InternalServerError:
      description: Terjadi kesalahan internal pada server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"